+++ internal/bot/components/misc/pressf.go
@@ -8,7 +8,6 @@ import (
 	"time"
 
 	"github.com/codemicro/lgballtDiscordBot/internal/bot/components/core"
-	"github.com/codemicro/lgballtDiscordBot/internal/config"
 	"github.com/codemicro/lgballtDiscordBot/internal/logging"
 	"github.com/codemicro/lgballtDiscordBot/internal/tools"
 	"github.com/skwair/harmony"
@@ -58,11 +57,7 @@ func newPressFTracker(bot *core.Bot, message *harmony.Message, duration time.Dur
 				logging.Error(err, "activePressF runner (nickname get)")
 				return
 			}
-			pronoun := "their"
-			if v.UserID == config.OwnerId {
-				pronoun = "her"
-			}
-			newText := fmt.Sprintf("%s\n**%s** has paid %s respects", message.Content, name, pronoun)
+			newText := fmt.Sprintf("%s\n**%s** has paid respects", message.Content, name)
 			_, err = apf.Bot.Client.Channel(apf.Message.ChannelID).EditMessage(context.Background(), apf.Message.ID, newText)
 			apf.Message.Content = newText
 			if err != nil {